- Implemented ConvertDollarToCent() in GameAnalyticsService.cs - Implementing rounding amount float parameter to int in GameAnalyticsService.RevenueEvent() - Changed type of amount parameter in IAnalyticsService.RevenueEvent() from int to float - Added Debug.Log() message for RevenueEvent() in FirebaseAnalyticsService.cs and GameAnalyticsService.cs and MetrixService.cs - Changed name of MetrixService.GetMetrixCurrencyCode() from currencyCode to currencyCodeNum

+++ narin-analytics-sdk_unity/Assets/NarinAnalyticsSDK/AnalyticsServices.cs
@@ -32,7 +32,7 @@ namespace Narin.Unity.Analytics {
             return ret;
         }
 
-        public void RevenueEvent(Currency currency, int amount, string itemType, string itemId, string cartType, string slug = null) {
+        public void RevenueEvent(Currency currency, float amount, string itemType, string itemId, string cartType, string slug = null) {
             foreach(var service in _services.Keys) {
                 _services[service].RevenueEvent(currency, amount, itemType, itemId, cartType, slug);
             }
